Room db module

+++ binance_api/src/main/java/dev/kokorev/binance_api/entity/BinanceExchangeInfoDTO.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class BinanceExchangeInfoDTO(
     @SerializedName("timezone") val timezone: String,
-    @SerializedName("serverTime") val serverTime: Int,
+    @SerializedName("serverTime") val serverTime: Long,
 //    @SerializedName("rateLimits") val rateLimits: ArrayList<RateLimit>,
     @SerializedName("exchangeFilters") val exchangeFilters: ArrayList<String>,
     @SerializedName("symbols") val binanceSymbolDTOS: ArrayList<BinanceSymbolDTO>
